# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GideonDB experimental API
  version: 0.0.1
  description: Phase 1 API; compatibility is not yet promised.
  contact:
    url: https://github.com/Bharanipbk/gideondb
servers:
  - url: http://127.0.0.1:6333/v1
security:
  - bearerAuth: []
paths:
...
  /collections:
    get:
      operationId: listCollections
      parameters:
        - {$ref: "#/components/parameters/PageLimit"}
        - {$ref: "#/components/parameters/PageCursor"}
      responses:
        "200":
          description: Collection configurations.
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    items: {$ref: "#/components/schemas/CollectionConfig"}
                  next_cursor: {type: string}
                required: [collections, next_cursor]
    post:
      operationId: createCollection
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CollectionConfig"}
      responses:
        "201": {description: Collection created.}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /collections/{name}:
    parameters:
      - {$ref: "#/components/parameters/CollectionName"}
    get:
      operationId: describeCollection
      responses:
        "200": {description: Collection description.}
        "404": {$ref: "#/components/responses/Error"}
    delete:
      operationId: deleteCollection
      responses:
        "204": {description: Collection deleted.}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    CollectionName:
      name: name
      in: path
      required: true
      schema: {type: string, pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$"}
...
    PageLimit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 100}
    PageCursor:
      name: cursor
      in: query
      schema: {type: string}
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    CollectionConfig:
      type: object
      additionalProperties: false
      properties:
        name: {type: string}
        dimension: {type: integer, minimum: 1}
        metric: {type: string, enum: [cosine, dot, l2]}
        shard_count: {type: integer, minimum: 1}
        index: {$ref: "#/components/schemas/IndexConfig"}
      required: [name, dimension, metric, shard_count]
    IndexConfig:
      type: object
      additionalProperties: false
      properties:
        type: {type: string, enum: [flat, hnsw], default: flat}
        m: {type: integer, minimum: 2, maximum: 128}
        ef_construction: {type: integer, minimum: 2, maximum: 10000}
        ef_search: {type: integer, minimum: 1, maximum: 10000}
...
    Error:
      type: object
      properties:
        code: {type: string}
        message: {type: string}
      required: [code, message]